moved copy to Util
