+++ app/src/main/java/ie/delilahsthings/soothingloop/Util.java
@@ -9,25 +9,21 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
 
 public abstract class Util {
-    public static String[] safe_list(File dir) {
-        String[] lst = dir.list();
-        if(lst==null){
-            return new String[0];
-        }
-        else {
-            return lst;
-        }
-    }
 
-    static void sleep(long milis)
-    {
-        try {
-            Thread.sleep(milis);
+    static int copy(InputStream in, OutputStream out) throws IOException {
+        // Transfer bytes from in to out
+        byte[] buf = new byte[1024];
+        int len, total=0;
+        while ((len = in.read(buf)) > 0) {
+            out.write(buf, 0, len);
+            total+=len;
         }
-        catch (InterruptedException e) {
-        }
+        return total;
     }
 
     public static <T> T getChildOfType(ViewGroup viewGroup, Class<T> type)
@@ -66,5 +62,24 @@ public abstract class Util {
             }
         }
         return result;
+    }
+
+    public static String[] safe_list(File dir) {
+        String[] lst = dir.list();
+        if(lst==null){
+            return new String[0];
+        }
+        else {
+            return lst;
+        }
+    }
+
+    static void sleep(long milis)
+    {
+        try {
+            Thread.sleep(milis);
+        }
+        catch (InterruptedException e) {
+        }
     }
 }
